dockerfile refactor
