+++ engine/internal/controller/engines/Readme.MD
@@ -1,0 +1,6 @@
+
+
+https://medium.com/@diogok/on-golang-static-binaries-cross-compiling-and-plugins-1aed33499671
+
+https://github.com/github/hub/issues/1818
+
