+++ engine/internal/controller/engines/manifest-siv2-ffmpeg.json
@@ -1,0 +1,14 @@
+{
+	"engineId": "8bdb0e3b-ff28-4f6e-a3ba-887bd06e6440",
+	"url": "https://github.com/veritone/engine-toolkit",
+	"category": "ingestion",
+	"engineMode": "stream",
+	"externalCalls": [],
+	"preferredInputFormat": "application/json",
+	"outputFormats": ["video/mp4"],
+	"clusterSize": "custom",
+	"customProfile": "stream-ingestor",
+	"initialConcurrency": 50,
+	"maxConcurrency": 50,
+        "releaseNotes": "RELEASE_NOTES"
+}
