+++ engine/internal/controller/engines/manifest-siv2-playback.json
@@ -1,0 +1,14 @@
+{
+	"engineId": "352556c7-de07-4d55-b33f-74b1cf237f25",
+	"url": "https://github.com/veritone/engine-toolkit",
+	"category": "ingestion",
+	"engineMode": "stream",
+	"externalCalls": [],
+	"preferredInputFormat": "application/json",
+	"outputFormats": ["video/mp4"],
+	"clusterSize": "custom",
+	"customProfile": "stream-ingestor",
+	"initialConcurrency": 50,
+	"maxConcurrency": 50,
+        "releaseNotes": "RELEASE_NOTES"
+}
