--- engine/vendorsave
@@ -1,145 +1,0 @@
-vendors:
-- path: github.com/Shopify/sarama
-  rev: 1358e9c6e61694cd61b2daae79f5aa4b8073c976
-- path: github.com/antihax/optional
-  rev: c3f0ba9c1a592b971d66b2787679af55b5c58f21
-- path: github.com/armon/circbuf
-  rev: 5111143e8da2e98b4ea6a8f32b9065ea1821c191
-- path: github.com/aviddiviner/go-murmur
-  rev: b9740d71e571c1f4ccb570b9bc7f352329d3e600
-- path: github.com/aws/aws-sdk-go
-  rev: 3576772d916b5db1cb5516b772bcdc362349a177
-- path: github.com/aws/aws-sdk-go-v2
-  rev: ca0c9eedb2220a76cb90bf7775b14debd771fd03
-- path: github.com/beorn7/perks
-  rev: 37c8de3658fcb183f997c4e13e8337516ab753e6
-- path: github.com/bsm/sarama-cluster
-  rev: d5779253526cc8a3129a0e5d7cc429f4b4473ab4
-- path: github.com/c2h5oh/datasize
-  rev: 4eba002a5eaea69cf8d235a388fc6b65ae68d2dd
-- path: github.com/davecgh/go-spew
-  rev: d8f796af33cc11cb798c1aaeb27a4ebc5099927d
-- path: github.com/eapache/go-resiliency
-  rev: b98ce2825f7276cc41027348986a241ef5d62fd8
-- path: github.com/eapache/go-xerial-snappy
-  rev: 776d5712da21bc4762676d614db1d8a64f4238b0
-- path: github.com/eapache/queue
-  rev: 093482f3f8ce946c05bcba64badd2c82369e084d
-- path: github.com/gabriel-vasile/mimetype
-  rev: 1f0a0f03d7687716bdf32add0e1c8061ea0a3f45
-- path: github.com/golang/protobuf
-  rev: 1680a479a2cfb3fa22b972af7e36d0a0fde47bf8
-- path: github.com/golang/snappy
-  rev: ff6b7dc882cf4cfba7ee0b9f7dcc1ac096c554aa
-- path: github.com/google/uuid
-  rev: c2e93f3ae59f2904160ceaab466009f965df46d6
-- path: github.com/gorilla/mux
-  rev: 884b5ffcbd3a11b730f0b75f5c86ac408753c34d
-- path: github.com/h2non/filetype
-  rev: 567fd48033ecc8ba8f04eb6d08eae427e6bf7376
-- path: github.com/hashicorp/go-uuid
-  rev: f32f5fe8d6a86bfd2a4f58998002f7b0b91cf78f
-- path: github.com/hpcloud/tail
-  rev: a1dbeea552b7c8df4b542c66073e393de198a800
-- path: github.com/jcmturner/gofork
-  rev: dc7c13fece037a4a36e2b3c69db4991498d30692
-- path: github.com/jmespath/go-jmespath
-  rev: c2b33e8439af944379acbdd9c3a5fe0bc44bd8a5
-- path: github.com/klauspost/compress
-  rev: 30993c63e1b02b47dd3878e74c8db50833706ecc
-- path: github.com/konsorten/go-windows-terminal-sequences
-  rev: f55edac94c9bbba5d6182a4be46d86a2c9b5b50e
-- path: github.com/machinebox/graphql
-  rev: 3a92531802258604bd12793465c2e28bc4b2fc85
-- path: github.com/matryer/is
-  rev: 7fc08349c0c760a7fe79d43c1939a7fc050a444b
-- path: github.com/matttproud/golang_protobuf_extensions
-  rev: c182affec369e30f25d3eb8cd8a478dee585ae7d
-- path: github.com/mitchellh/mapstructure
-  rev: 3536a929edddb9a5b34bd6861dc4a9647cb459fe
-- path: github.com/nfnt/resize
-  rev: 83c6a9932646f83e3267f353373d47347b6036b2
-- path: github.com/opentracing/opentracing-go
-  rev: a7454ce5950e83caac9774ef32cc8979e5975f30
-- path: github.com/pborman/uuid
-  rev: 8b1b92947f46224e3b97bb1a3a5b0382be00d31e
-- path: github.com/pierrec/lz4
-  rev: 0554902aa74a6a41b367c6352ebcc922b26782b1
-- path: github.com/pkg/errors
-  rev: 27936f6d90f9c8e1145f11ed52ffffbfdb9e0af7
-- path: github.com/pmezard/go-difflib
-  rev: 5d4384ee4fb2527b0a1256a821ebfc92f91efefc
-- path: github.com/prometheus/client_golang
-  rev: eb4edc44747f2b520aa8123453ba0618905b92f9
-- path: github.com/prometheus/client_model
-  rev: 14fe0d1b01d4d5fc031dd4bec1823bd3ebbe8016
-- path: github.com/prometheus/common
-  rev: 287d3e634a1e550c9e463dd7e5a75a422c614505
-- path: github.com/prometheus/procfs
-  rev: 648defec548e893f87609e8a3daa1d82e8981221
-- path: github.com/rcrowley/go-metrics
-  rev: cac0b30c2563378d434b5af411844adff8e32960
-- path: github.com/rwcarlsen/goexif
-  rev: 9e8deecbddbd4989a3e8d003684b783412b41e7a
-- path: github.com/satori/go.uuid
-  rev: b2ce2384e17bbe0c6d34077efa39dbab3e09123b
-- path: github.com/segmentio/kafka-go
-  rev: 5e37e0483bf9aac5ec283e3a76086e0289952b97
-- path: github.com/sirupsen/logrus
-  rev: fb62dbe2f2a2c88b150fa1668a773505a7920a3d
-- path: github.com/stretchr/objx
-  rev: 35313a95ee26395aa17d366c71a2ccf788fa69b6
-- path: github.com/stretchr/testify
-  rev: 85f2b59c4459e5bf57488796be8c3667cb8246d6
-- path: github.com/uber/jaeger-client-go
-  rev: d354d6ac8b02fb0570ce9cc86c9384f40df0bf6c
-- path: github.com/uber/jaeger-lib
-  rev: a87ae9d84fb038a8d79266298970720be7c80fcd
-- path: github.com/veritone/core-messages
-  rev: 8bc2cb14ab53f45d9c794e08a02993f6511dea06
-- path: github.com/veritone/edge-messages
-  rev: 56ba8e76daaa6148acde1ced17c9961e33eec7b0
-- path: github.com/veritone/edge-stream-ingestor
-  rev: 04c9125c03d344fe02cbcd2906bc9b4f521b7196
-- path: github.com/veritone/go-messaging-lib
-  rev: 74a8be2dabf938d92c88c892cd80e3520446f9bc
-- path: github.com/veritone/graphql
-  rev: fa96d6ddd0426485f5bb23575132e7f703a8b37a
-- path: github.com/veritone/realtime
-  rev: 686a3f0ff611456feb4bfb2f9f8f7aaab16ae0e1
-- path: github.com/veritone/webstream-adapter
-  rev: 53ce102647f497de34d3125f0ac823d390f20a4e
-- path: go.uber.org/atomic
-  rev: ef0d20d85b01f4708a8085bca571838bf59a3f77
-- path: go.uber.org/multierr
-  rev: 71e610a0e48dbda460d9c18adca8b5f2de04a7c1
-- path: go.uber.org/zap
-  rev: d8445f34b4ae69130c5337a6fb65f317e3b71f76
-- path: golang.org/x/crypto
-  rev: 87dc89f01550277dc22b74ffcf4cd89fa2f40f4c
-- path: golang.org/x/net
-  rev: da9a3fd4c5820e74b24a6cb7fb438dc9b0dd377c
-- path: golang.org/x/oauth2
-  rev: 0f29369cfe4552d0e4bcddc57cc75f4d7e672a33
-- path: golang.org/x/sys
-  rev: b09406accb4736d857a32bf9444cd7edae2ffa79
-- path: golang.org/x/text
-  rev: 3d0f7978add91030e5e8976ff65ccdd828286cba
-- path: google.golang.org/appengine
-  rev: 16bce7d3dc4e458f2f6f56a1349cbbfcdc8a8fdf
-- path: gopkg.in/fsnotify.v1
-  rev: c2828203cd70a50dcccfb2761f8b1f8ceef9a8e9
-- path: gopkg.in/jcmturner/aescts.v1
-  rev: f6abebb3171c4c1b1fea279cb7c7325020a26290
-- path: gopkg.in/jcmturner/dnsutils.v1
-  rev: 13eeb8d49ffb74d7a75784c35e4d900607a3943c
-- path: gopkg.in/jcmturner/gokrb5.v7
-  rev: 363118e62befa8a14ff01031c025026077fe5d6d
-- path: gopkg.in/jcmturner/rpc.v1
-  rev: 99a8ce2fbf8b8087b6ed12a37c61b10f04070043
-- path: gopkg.in/tomb.v1
-  rev: dd632973f1e7218eb1089048e0798ec9ae7dceb8
-- path: gopkg.in/yaml.v2
-  rev: f221b8435cfb71e54062f6c6e99e9ade30b124d5
-- path: github.com/go-redis/redis
-  rev: 3ad29557285e162b009c3e00abf1e68599e61d04
